Move the rest of global variables
